Extract 'ObjectMetadata' class from 'HashStore' to 'FileHashStore', revise and organize affected code
