+++ src/hashstore/filehashstore.py
@@ -9,12 +9,13 @@ import hashlib
 import os
 import logging
 import inspect
+import fcntl
+import yaml
+from dataclasses import dataclass
 from pathlib import Path
 from contextlib import closing
 from tempfile import NamedTemporaryFile
-import fcntl
-import yaml
-from hashstore import HashStore, ObjectMetadata
+from hashstore import HashStore
 from hashstore.filehashstore_exceptions import (
     CidRefsContentError,
     CidRefsDoesNotExist,
@@ -2611,3 +2612,25 @@ class Stream(object):
             self._obj.close()
         else:
             self._obj.seek(self._pos)
+
+
+@dataclass
+class ObjectMetadata:
+    """Represents metadata associated with an object.
+
+    The `ObjectMetadata` class represents metadata associated with an object, including
+    a persistent or authority-based identifier (`pid`), a content identifier (`cid`),
+    the size of the object in bytes (`obj_size`), and an optional list of hex digests
+    (`hex_digests`) to assist with validating objects.
+
+    :param str pid: An authority-based or persistent identifier
+    :param str cid: A unique identifier for the object (Hash ID, hex digest).
+    :param int obj_size: The size of the object in bytes.
+    :param list hex_digests: A list of hex digests to validate objects
+        (md5, sha1, sha256, sha384, sha512) (optional).
+    """
+
+    pid: str
+    cid: str
+    obj_size: int
+    hex_digests: dict
